formatting and refactoring
